Fabric_assigment/sdk/gateway/fabric_gateway.go
package main

import (
	"fmt"

	"github.com/hyperledger/fabric-sdk -go/pkg/common/providers/fabric"
	"github.com/hyperledger/fabric-sdk-go/pkg/fabsdk"
)

type FabricGateway struct{}

func (fg *FabricGateway) CreateAccount(dealerID string, msisdn string, mpin string, balance int, status string, transAmount int, transType string, remarks string) error {
	fmt.Println("Creating account through gateway...")
	// Implement account creation logic here
	return nil
}

func (fg *FabricGateway) UpdateAccount(dealerID string, balance int, transAmount int, transType string, remarks string) error {
	fmt.Println("Updating account through gateway...")
	// Implement account update logic here
	return nil
}

func (fg *FabricGateway) GetAccount(dealerID string) ([]byte, error) {
	fmt.Println("Getting account through gateway...")
	// Implement account retrieval logic here
	return nil, nil
}

func (fg *FabricGateway) GetHistory(dealerID string) ([]byte, error) {
	fmt.Println("Getting history through gateway...")
	// Implement history retrieval logic here
	return nil, nil
}